test: updated tests
