+++ sandbox/api/tests/test_app.py
@@ -103,44 +103,6 @@ def test_related_person(
 
 
 @pytest.mark.parametrize(
-    "request_args,response_file_name,status_code",
-    [
-        (
-            "identifier=9000000041",
-            "./api/examples/errors/not-found.yaml",
-            404,
-        ),
-        (
-            "identifier=9000000017&patient:identifier=9000000041",
-            "./api/examples/errors/not-found.yaml",
-            404,
-        ),
-    ],
-)
-@patch(f"{UTILS_FILE_PATH}.generate_response_from_example")
-def test_related_person__not_found(
-    mock_generate_response_from_example: MagicMock,
-    request_args: str,
-    response_file_name: str,
-    status_code: int,
-    client: object,
-) -> None:
-    """Test related_persons endpoint."""
-    # Arrange
-    mock_generate_response_from_example.return_value = mocked_response = Response(
-        dumps({"data": "mocked"}), status=status_code, content_type="application/json"
-    )
-    # Act
-    response = client.get(f"{RELATED_PERSON_API_ENDPOINT}?{request_args}")
-    # Assert
-    mock_generate_response_from_example.assert_called_once_with(
-        response_file_name, status_code
-    )
-    assert response.status_code == status_code
-    assert response.json == json.loads(mocked_response.get_data(as_text=True))
-
-
-@pytest.mark.parametrize(
     "url_path,response_file_name,status_code",
     [
         (
@@ -173,69 +135,24 @@ def test_questionnaire_response(
     ("request_args,response_file_name,status_code"),
     [
         (
-            "performer:identifier=9000000017&status=active",
-            "./api/examples/GET_Consent/filtered-relationships-status-active.yaml",
+            "performer:identifier=9000000025",  # No performer record found error
+            "./api/examples/GET_Consent/no-relationships.yaml",
             200,
         ),
         (
-            "performer:identifier=9000000017&status=inactive",
-            "./api/examples/GET_Consent/filtered-relationships-status-inactive.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000017&status=proposed,active",
-            "./api/examples/GET_Consent/filtered-relationships-status-proposed-active.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000022",
-            "./api/examples/GET_Consent/multiple-relationships.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000022&_include=Consent:patient",
-            "./api/examples/GET_Consent/multiple-relationships-include-patient.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000022&_include=Consent:performer",
-            "./api/examples/GET_Consent/multiple-relationships-include-performer.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000022&_include=Consent:performer,Consent:patient",
-            "./api/examples/GET_Consent/multiple-relationships-include-performer-patient.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000010",
-            "./api/examples/GET_Consent/single-consenting-adult-relationship.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000010&_include=Consent:performer,Consent:patient",
-            "./api/examples/GET_Consent/single-consenting-adult-relationship-include-performer-patient.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000019",
-            "./api/examples/GET_Consent/single-mother-child-relationship.yaml",
-            200,
-        ),
-        (
-            "performer:identifier=9000000019&_include=Consent:performer,Consent:patient",
-            "./api/examples/GET_Consent/single-mother-child-relationship-include-performer-patient.yaml",
-            200,
+            "performer:identifier=9000000999",  # No performer record found error
+            "./api/examples/errors/not-found.yaml",
+            404,
         ),
     ],
 )
-@patch(f"{UTILS_FILE_PATH}.generate_response_from_example")
-def test_consent(
+@patch(f"{APP_FILE_PATH}.generate_response_from_example")
+def test_consent_from_app(
     mock_generate_response_from_example: MagicMock,
     request_args: str,
     response_file_name: str,
     status_code: int,
-    client: object,
+    client: object
 ) -> None:
     """Test Consent endpoint."""
     mock_generate_response_from_example.return_value = mocked_response = Response(
@@ -249,41 +166,6 @@ def test_consent(
     )
     assert response.status_code == status_code
     assert response.json == json.loads(mocked_response.get_data(as_text=True))
-
-
-@patch(f"{APP_FILE_PATH}.generate_response_from_example")
-def test_consent_no_relationships(
-    mock_generate_response_from_example: MagicMock, client: object
-) -> None:
-    """Test Consent endpoint."""
-    mock_generate_response_from_example.return_value = mocked_response = Response(
-        dumps({"data": "mocked"}), status=200, content_type="application/json"
-    )
-    # Act
-    response = client.get(f"{CONSENT_API_ENDPOINT}?performer:identifier=9000000025")
-    # Assert
-    mock_generate_response_from_example.assert_called_once_with(
-        "./api/examples/GET_Consent/no-relationships.yaml", 200
-    )
-    assert response.status_code == 200
-    assert response.json == json.loads(mocked_response.get_data(as_text=True))
-
-
-@patch(f"{APP_FILE_PATH}.generate_response_from_example")
-def test_consent__404_bad_request(
-    mock_generate_response_from_example: MagicMock,
-    client: object,
-) -> None:
-    """Test Consent endpoint."""
-    mock_generate_response_from_example.return_value = Response(
-        dumps({"data": "mocked"}), status=404, content_type="application/json"
-    )
-    # Act
-    client.get(f"{CONSENT_API_ENDPOINT}?performer:identifier=9000000999")
-    # Assert
-    mock_generate_response_from_example.assert_called_once_with(
-        "./api/examples/errors/not-found.yaml", 404
-    )
 
 
 @patch(f"{APP_FILE_PATH}.remove_system")
