+++ specification/examples/responses/errors/invalid-identifier-system.yaml
@@ -1,0 +1,16 @@
+InvalidIdentifierSystemBundle:
+  summary: Bad request identifier system invalid
+  description: 400 error response bundle for an invalid system of the identifier
+  value:
+    issue:
+      - code: invalid
+        details:
+          coding:
+            - code: INVALID_IDENTIFIER_SYSTEM
+              display: Invalid identifier system.
+              system: >-
+                https://fhir.nhs.uk/R4/CodeSystem/ValidatedRelationships-ErrorOrWarningCode
+              version: '1'
+        diagnostics: The identifier system is not valid.
+        severity: error
+    resourceType: OperationOutcome
